ANDROID: kvm/vfio: Add pviommu group and attach/get attrs
pvIOMMU would configure a paravirtualised IOMMU exposed to guests
that is controlled by the hypervisor.

This pvIOMMU would be used with VFIO assigned devices. As this
requires both KVM and VFIO interaction, it was added in KVM-VFIO
device.

This patch adds pviommu_attach operation which attaches a pvIOMMU
to a VM.

pviommu_attach would use the KVM VM for this device and return a
fd for the guest pvIOMMU.

It also adds attribute KVM_DEV_VFIO_PVIOMMU_GET_INFO to probe info about
iommu for a VFIO device fd, for now only number of SIDs is returned.
This can be used by the VMM to then configure the virtual SID (vSID) for
the IOMMU IDs.

Bug: 357781595
Bug: 348382247
Bug: 236685427
Bug: 458241298
Change-Id: I030f04e84fa8fd1c54c2d035f5c09f021f648d22
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 89458eb2b18369956916b6e3ca7ecf9731081359)