ANDROID: KVM: arm64: iommu: Add context for guest teardown
As memory management for host and guests are different we rely on
__get_vcpu to know which methods to use, however in some contexts
as in free_domain/block_dev, the guest will not be loaded and we'd
need the memory reclaim to be done through the correct pool.
To fix this we add a static per cpu variable that can be set as a
fallback if vcpu is not loaded to check if we are in a teardown
context.

Bug: 357781595
Bug: 348382247
Bug: 236685427
Bug: 458241298
Change-Id: I790d92a8dd6112bf40302bc0d6ec10582362db65
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 7db797ed3fbf012317b8f85503d745387563681f)