ANDROID: KVM: arm64: Drop struct pkvm_mapping from KMI
struct pkvm_mapping is currently exposed in the KMI via the pKVM module
interface as it's a member of the hyp_memcache. But the pkvm_mapping is
only used in the page-fault path at EL1, and no module should ever need
to use it. So let's turn this pointer into a void * to hide struct
pkvm_mapping from the KMI.

Bug: 357781595
Change-Id: I2fe5132a74b64405fa15805c0b346afcde94fc56
Signed-off-by: Quentin Perret <qperret@google.com>