ANDROID: KVM: arm64: Support power domains
Unfortunately, as power management is not widely standardized we have
to work around that.

One implementation we can support is HOST_HVC, where the host
is in control of the power management and it notifies the
hypervisor about the updates.

This adds extra constraints to the IOMMUs, as they must reset to
blocking DMA traffic to be able to use this PD interface.

Unfortunately again, for SMMUv3 which is the only IOMMU currently
supported, there is no architectural way to discover this, so we
rely on enabling this driver when it fits the constraints, also
the driver sets GBPA and assumes that the SMMU retains across
power cycling.

Bug: 461757236
Bug: 458241298
Change-Id: Idf77a724b518d2e3133af159dfcfb32d47a3aa4d
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 7681c1e59bc0da6ef30f2a8f0d51c9278afb1567)