ANDROID: KVM: arm64: Fix pins for split requests on last-level mappings
For robustness, it is important to treat a split request as valid even
though the region is already mapped up to the last-level. However a
missing unpin call in that path means we could leak pins after the VM is
tear downed. Fix it.

Bug: 419548963
Bug: 357781595
Bug: 278011447
Change-Id: I2c0d69b2168135722b46d97358f46aff56ecac03
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>
(cherry picked from commit 6fbbc172dcfeffef4b09bea5866381f83fa99f3b)