ANDROID: iommu/io-pgtable-arm: Add support for IO_PGTABLE_QUIRK_IDMAP
Add a new quirk to allocate a page table which is identity mapped,
this means that IOVA must be equal to PADDR. It is only supported
by stage-2 page tables.

In this patch, this is used by the hypervisor to choose which memory
pool to allocate page tables from.
In the next patches, this can be used for:
- Coalescing support
- split blk unmap
- Extra assertions

Bug: 461757236
Bug: 458241298
Change-Id: I98d7b141e60290906468595d52ba3073d85c108e
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 22857575fda757fd85597ea0d0bbf9f3158e1bc2)