KVM: arm64: Trap access to ALLINT if FEAT_NMI not supported by the guest
Access to ALLINT is part of FEAT_NMI. If a guest vCPU does not support
this feature, any access to this register must be trapped to the
hypervisor (EL2).

KVM didn't configure this trap, potentially allowing a guest to toggle
all interrupt mask when it doesn't support FEAT_NMI. Fix this by
checking if the guest has FEAT_NMI support.

Signed-off-by: Fuad Tabba <tabba@google.com>