ANDROID: KVM: arm64: Remove token from pKVM module registration path
The token is populated on module registration to be passed for later
registration of a module HVC. Hidden was the base address of the module
in the hyp private mapping. This address is now stored in the
pkvm_el2_module struct and doesn't need to be passed over to the caller.
Remove it completely.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>