ANDROID: arm64: Disable MTE at EL1 and EL0 when not supported or disabled
If MTE is not supported by the hardware, or disabled in the kernel
configuration or in the command line. it should be disabled. If hardware
supports MTE, but it is disabled by software, this could result in a
buggy or a malicious host using the Memory Tag Extention, crashing the
system, or compromising guests.

Disable Memory Tagging altogether by trapping it to EL2 if it's not
supported or disabled at the host.

Bug: 458241298
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: If539d9366e5c06155f63d6aea9bef123c8bc251d
(cherry picked from commit f3c79f686aaa4b4cbeb8992048b255d2ffa645a8)