ANDROID: KVM: arm64: Disable Memory Tagging for all guests if not supported
Currently, pKVM trusts the host for the state of HCR_EL2, which controls
Memory Tagging, for non-protected guests. However, if Memory Tagging is
disabled for the host, a buggy or a malicious host could enable it for
its (non-protected) guests.

Avoid this from happening by always disabling it for all guests when
activating traps.

Note: An upcoming patch going upstream will solve this by ensuring that
HCR_ATA (which enabled Memory Tagging) is not set to begin with. For
now, let's use the more aggressive approach to ensure we don't miss
anything. Drop this patch once the upstream one is in.

Bug: 458241298
Signed-off-by: Fuad Tabba <tabba@google.com>
Change-Id: I05ea9655ceadcbdf46ca9f63d8dcecd333d3e91c
(cherry picked from commit 766be88813859dd0533d4dbbad18dbf041d7b32e)