FROMLIST: iommu/io-pgtable-arm: Factor kernel specific code out
Some of the used APIs are only part of the kernel and are not
available in the hypervisor, factor those out:
- alloc/free memory
- CMOs
- virt/phys conversions

Which are implemented by the kernel in io-pgtable-arm-kernel.c and
similarly for the hypervisor later in this series.

va/pa conversion kept as macros.

Link: https://patch.msgid.link/20251117184815.1027271-5-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: Iefc45534aa4b6c22a3987f4a9901764131a9062d
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 3f62fdaf0e171a91e2559431cf4f3f554856545b)