FROMLIST: KVM: arm64: iommu: Shadow host stage-2 page table
Create a shadow page table for the IOMMU that shadows the
host CPU stage-2 into the IOMMUs to establish DMA isolation.

An initial snapshot is created after the driver init, then
on every permission change a callback would be called for
the IOMMU driver to update the page table.

For some cases, an SMMUv3 may be able to share the same page
table used with the host CPU stage-2 directly.
However, this is too strict and requires changes to the core hypervisor
page table code, plus it would require the hypervisor to handle IOMMU
page faults. This can be added later as an optimization for SMMUV3.

Link: https://patch.msgid.link/20251117184815.1027271-10-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: Ifccc90dc3759976bf1c2104e44f60cc04b41694c
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 2f085b2a7f78ef8d0010c006e45cb6a34cd78751)