FROMLIST: KVM: arm64: iommu: Support DABT for IOMMU
SMMUv3 driver need to trap and emulate access to the MMIO space
to provide emulation for the kernel.

Add a handler for DABTs for IOMMU drivers to be able to do so, in
case the host fault in page, check if it's part of IOMMU emulation
first.

Link: https://patch.msgid.link/20251117184815.1027271-12-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: I786835ef422addd53064d1732d5684f4505ae964
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 825a1f1f6c11c972e1f8f7031c4da1adc4c0d7a1)