FROMLIST: iommu/arm-smmu-v3: Support probing KVM emulated devices
When KVM runs in protected mode, and CONFIG_ARM_SMMU_V3_PKVM
is enabled, it will manage the SMMUv3 HW using trap and emulate
and present emulated SMMUs to the host kernel.

In that case, those SMMUs will be on the aux bus, so make it
possibly to the driver to probe those devices.
Otherwise, everything else is the same as the KVM emulation
complies with the architecutre, so the driver doesn't need
to be modified.

Suggested-by: Jason Gunthorpe <jgg@ziepe.ca>
Link: https://patch.msgid.link/20251117184815.1027271-15-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: I7e7ef338ef02d80870023c91250c593e0e2589ad
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit b8b3d78343d038b9ece81700299372ee20571f07)