FROMLIST: iommu/arm-smmu-v3-kvm: Shadow stream table
This patch allocates the shadow stream table per SMMU.
We choose the size of that table to be 1MB which is the
max size used by host in the case of 2 levels.

In this patch all the host writes are still paththrough for
bisectibility, that is changed next where CFGI commands will be
trapped and used to update the shadow copy hypervisor that
will be used by HW.

Similar to the command queue, the host stream table is
shared/unshared each time the SMMU is enabled/disabled.

Handling of L2 tables is also done in the next patch when
the shadowing is added.

Link: https://patch.msgid.link/20251117184815.1027271-23-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: I7cc2ce60f80a0fbdbbf4ff29db87e6650b5f40ab
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 39ccbe3f937451432bd4cf493b050932baa785d4)