FROMLIST: iommu/arm-smmu-v3-kvm: Shadow STEs
This patch adds STE emulation, this is done when the host sends the
CFGI_STE command.

In this patch we copy the STE as is to the shadow owned by the hypervisor,
in the next patch, stage-2 page table will be attached.

Link: https://patch.msgid.link/20251117184815.1027271-24-smostafa@google.com

Bug: 461757236
Bug: 458241298
Change-Id: I0048f6244cd1acb06ed029ac09d7e925a584612f
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit fca4717251edd4bd2ec0af5ca5b80a1684dfac39)