ANDROID: iommu/pkvm-pviommu: Set max_pasids
The IOMMU core needs to know the max number of supported
pasids, this information is not exposed to the guest driver.

Although it is possible to do, that will require an new guest
HVC OP and new ops in EL2 to route to the driver, which is
intrusive for this frozen kernel.

As the core code only needs this information to do checks that
a device pasid is within range, and that checks will be again
done at EL2 from the IOMMU driver, we can just export a large
range to the core IOMMU.
And if a guest driver is misbehaving it will just end with a
failed attach call but from the EL1 driver instead of the
core code.
Choose this value to be 256 which is the actual MAX for the
SMMUv3 driver; the only driver supporting pasids in tree.

Bug: 448427307
Change-Id: I19b3e486551e8f2aecc98377f0f694f9cae3a76e
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 401c8883eb8e789cda2473efb03f4b787d284160)