ANDROID: KVM: arm64: Don't use S2FWB for VM with devices
At the moment, VFIO doesn't allow devices that are not
"dma-coherent" to be assigned to VMs.
However, that check doesn't capture the full topology,
KVM relies on the IOMMU cap "IOMMU_CAP_ENFORCE_CACHE_COHERENCY",
to determine if it's possible to have non-coherent access from
the guest VM, which calls "kvm_arch_register_noncoherent_dma".
On x86, that will change KVM handling of guest page tables to honor
PAT's setting. And it's not implemented on Arm64.

On Arm64, the guest VM can already map memory as non-coherent in
the CPU, and in the absence of FWB, the hypervisor already has
CMOs to cover this case, all we need to do is disable FWB for
guests if it has devices.

We rely on the hypervisor tracking rather than the host notifying
the hypervisor through kvm_arch_register_noncoherent_dma() as it's
not trusted, plus this will always be true as the IOMMU implementation
is PCI related.

One intersting case is that __clean_dcache_guest_page() ellides CMO
if FWB is exisiting system-wise on the assumption that the FWB is
always used in supported which is not true any more.

One solution require intrusive re-architecutre to pass and propagate
the pgt in dcache_clean_inval_poc().

However, it seems we rely on the page table only calling those if it
doesn't have FWB, there was one missing case in the map walker.

And we can inline on extra check in pkvm_pgtable_stage2_flush()

Bug: 463588171
Bug: 448427307
Bug: 368611573
Change-Id: Iffe14e5db89a0c1ff115090b9338bca6eddfd702
Signed-off-by: Mostafa Saleh <smostafa@google.com>
(cherry picked from commit 70904a635c2beaf5c04c5583b7e9e918adda1e0a)